refac: passage SF7

+++ src/Command/ClearOldNotificationCommand.php
@@ -1,10 +1,10 @@
 <?php
 /*
- * Copyright (c) 2022. | David Annebicque | IUT de Troyes  - All Rights Reserved
+ * Copyright (c) 2024. | David Annebicque | IUT de Troyes  - All Rights Reserved
  * @file /Users/davidannebicque/Sites/intranetV3/src/Command/ClearOldNotificationCommand.php
  * @author davidannebicque
  * @project intranetV3
- * @lastUpdate 16/12/2022 12:09
+ * @lastUpdate 23/02/2024 21:40
  */
 
 /*
@@ -46,6 +46,6 @@ class ClearOldNotificationCommand extends Command
 
         $io->success(sprintf('"%d" notifications supprimées.', $count));
 
-        return (int)Command::SUCCESS;
+        return Command::SUCCESS;
     }
 }
