+++ src/Command/UpdateVilleCommand.php
@@ -1,10 +1,10 @@
 <?php
 /*
- * Copyright (c) 2022. | David Annebicque | IUT de Troyes  - All Rights Reserved
+ * Copyright (c) 2024. | David Annebicque | IUT de Troyes  - All Rights Reserved
  * @file /Users/davidannebicque/Sites/intranetV3/src/Command/UpdateVilleCommand.php
  * @author davidannebicque
  * @project intranetV3
- * @lastUpdate 16/12/2022 12:09
+ * @lastUpdate 23/02/2024 21:40
  */
 
 /*
@@ -67,6 +67,6 @@ class UpdateVilleCommand extends Command
 
         $io->success(sprintf('%d Villes mises à jour.', $i));
 
-        return (int)Command::SUCCESS;
+        return Command::SUCCESS;
     }
 }
